+++ eyeWriterTracker/src/Tracking/glintFinder.h
@@ -16,33 +16,33 @@
 #include "ofxVectorMath.h"
 
 class glintFinder {
-	
+
 public:
-	
+
 	void		setup(int bigEyeWidth, int bigEyeHeight, float magRatio);
-	bool		update(ofxCvGrayscaleAdvanced & blackEyeImg, float threshold, float minBlobSize, float maxBlobSize);
+	bool		update(ofxCvGrayscaleAdvanced & blackEyeImg, int threshold, int minBlobSize, int maxBlobSize);
 	ofPoint		getGlintPosition(bool bIsLeftOrTop);
-	
+
 	ofxCvContourFinder		contourFinder;
 	ofxCvGrayscaleImage		eyeImage;
-	
+
 	float				subtractVal;
 	float				multiplyVal;
 
 	float				w,h;
 	float				magRatio;
-	
+
 	ofPoint				leftOrTopGlintPos;
 	ofPoint				rightOrBottomGlintPos;
-	
+
 	int					biggestID;
 	int					smallestID;
-	
+
 	bool				bFound;
-	
+
 	ofRectangle			glintROI;
 	bool				bIsVerticalLED;
-	
+
 };
 
 #endif /*_GLINT_FINDER_*/
